% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dense.fd.R
\name{dense.fd}
\alias{dense.fd}
\title{Generate functional data on a dense grid}
\usage{
dense.fd(mu, X, n, m, sig0 = NULL, snr = 5, domain = c(0, 1),
  grid = seq(domain[1], domain[2], length.out = m))
}
\arguments{
\item{mu}{a function, scalar or a vector defining the mean function}

\item{X}{the centered stochastic process, a function that
X(tObs,n) return n*len(tObs) matrix with n trajectories}

\item{n}{sample size}

\item{m}{sampling rate, ignored if \code{grid} is specified}

\item{sig0}{the std of measurement errors, if NULL, determined by snr}

\item{snr}{the signal to noise ratio to determine sig0}

\item{domain}{the domain}

\item{grid}{the grid on which the trajectories are observed}
}
\value{
a \code{n*m} matrix with the following attributes
     \item{sig0}
     \item{snr}{signal-to-nois ratio}
     \item{uncontaminated}{n*m matrix of clean observations}
     \item{grid}{a grid of points in \code{domain}}
     \item{domain}{the domain}
}
\description{
Generate functional data on a dense grid
}
\examples{
Y <- dense.fd(mu=1, X=gaussian.process(), n=10, m=20)
Y <- dense.fd(mu=cos, X=kl.process(),n=100, m=20)
}
