% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mul.data.R
\name{gen.mul.data}
\alias{gen.mul.data}
\title{Generate multivariate data with the given mean vector and covariance matrix}
\usage{
gen.mul.data(
  n = 100,
  K = 2,
  mu = rep(0, K),
  Sigma = diag(K),
  distribution = c("GAUSSIAN", "LAPLACE", "EXPONENTIAL")
)
}
\arguments{
\item{n}{sample size}

\item{K}{dimension}

\item{mu}{mean vector}

\item{Sigma}{covariance matrix}

\item{distribution}{distribution}
}
\value{
a \code{n*K} matrix
}
\description{
Generate multivariate data with the given mean vector and covariance matrix
}
\keyword{internal}
