% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reg.fd.R
\name{plot.dense.fd}
\alias{plot.dense.fd}
\title{Plot Regular Functional Data}
\usage{
\method{plot}{dense.fd}(x, ...)
}
\arguments{
\item{x}{the data object generated from \code{\link{reg.fd}}.}

\item{...}{other parameters passed to \code{matplot}.}
}
\value{
a plot of the dataset \code{x}.
}
\description{
Plot Regular Functional Data
}
