% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse.fd.R
\name{plot.sparse.fd}
\alias{plot.sparse.fd}
\title{plot sparse functional data}
\usage{
\method{plot}{sparse.fd}(x, ...)
}
\arguments{
\item{x}{the object generated by \code{sparse.fd}}

\item{...}{other parameters passed to \code{plot} and \code{lines}}
}
\description{
plot sparse functional data
}
