% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{regular.grid}
\alias{regular.grid}
\title{Generate a Regular Grid}
\usage{
regular.grid(M = 100, domain = c(0, 1), h = 1/(2 * M))
}
\arguments{
\item{M}{the number of points}

\item{domain}{the interval}

\item{h}{margin at boundaries}
}
\value{
equally spaced \code{M} points in \code{domain}.
}
\description{
Generate a Regular Grid
}
