% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{rep.col}
\alias{rep.col}
\title{replicate a column vector into a matrix}
\usage{
\method{rep}{col}(x, n)
}
\arguments{
\item{x}{the column vector to be replicated}

\item{n}{replicate the column vector n times}
}
\value{
a matrix of the dimension of \code{length(x)} by \code{n}
}
\description{
replicate a column vector into a matrix
}
\keyword{internal}
