% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse.fd.R
\name{sparse.fd}
\alias{sparse.fd}
\title{Generate sparse functional data}
\usage{
sparse.fd(mu, X, n, m, sig = NULL, snr = 5, domain = c(0, 1), delta = 1)
}
\arguments{
\item{mu}{a function or scalar defining the mean function}

\item{X}{the centered stochastic process, a function that
X(tObs,n) return n*len(tObs) matrix with n trajectories}

\item{n}{sample size}

\item{m}{sampling rate per trajectory. If \code{m} is an integer, than m_i is sampled from 1+Poison(m-1). If \code{m} is a vector of integers, m_i is sampled uniformly from this vector. If \code{m} is a function, it is then assumed to be a function f such that f(k) randomly generates k positive integers.}

\item{sig}{the std of measurement errors, if NULL, determined by snr}

\item{snr}{the signal to noise ratio to determine sig}

\item{domain}{the domain}

\item{delta}{the proportion of the domain to be observed. It must be a real number between 0 (exclusive) and 1 (inclusive). This argument is used to generate snippets. The onset time Oi is uniformly sampled from the interval [\code{domain[1]},\code{domain[2]-delta*(domain[2]-domain[1])}], and \eqn{t_{ij}}{tij} is uniformly sampled from [Oi,Oi+\code{delta*(domain[2]-domain[1])}].}
}
\value{
a list of two members
  \item{t}{a list of \code{n} items, each being a nondecreasing vector that represents the time points of observations for a subject}
  \item{y}{a list of \code{n} items, each being a vector representing the observations corresponding to \code{t}}
  and the following attributes
     \item{sig}{}
     \item{snr}{signal-to-nois ratio}
     \item{y0}{clean version of \code{y}}
     \item{domain}{the domain}
     \item{delta}{}
}
\description{
Generate sparse functional data
}
\examples{
# generate Gaussian samples with constant mean function 1
Y <- sparse.fd(mu=1, X=gaussian.process(), n=10, m=5)

# generate samples froma a process defined via K-L representation
Y <- sparse.fd(mu=cos, X=kl.process(eigen.values=1/(2^(1:25)),eigen.functions='FOURIER',distribution='LAPLACE'),n=10, m=5)

# generate samples with mi sample from 2,4,6
Y <- sparse.fd(mu=1, X=gaussian.process(cov=matern), n=10, m=c(2,4,6))
}
\references{
\insertRef{Lin2020}{synfd}
}
