% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processes.R
\name{white.noise}
\alias{white.noise}
\title{create a white noise process objecct}
\usage{
white.noise(sig = 1)
}
\arguments{
\item{sigma}{the sigma parameter of the white noise process}
}
\value{
a function hanlde in the form of \code{X(tObs,n)} which generates \code{n} independent Wiener trajectories observed at tObs
}
\description{
create a white noise process objecct
}
\examples{
X <- white.noise()
X(regular.grid(50),25)
}
