% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processes.R
\name{wiener.process}
\alias{wiener.process}
\title{create a Wiener process objecct}
\usage{
wiener.process(dispersion = 1)
}
\arguments{
\item{dispersion}{the dispersion parameter of the Wiener process}
}
\value{
a function hanlde in the form of \code{X(tObs,n)} which generates \code{n} independent Wiener trajectories observed at tObs
}
\description{
create a Wiener process objecct
}
\examples{
X <- wiener.process()
X(regular.grid(50),25)
}
