% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processes.R
\name{centered.process}
\alias{centered.process}
\title{create a centered random process objecct}
\usage{
centered.process(name = c("WIENER", "WHITE.NOISE", "KL",
  "KARHUNEN.LOEVE", "GAUSSIAN"), domain = c(0, 1), ...)
}
\arguments{
\item{name}{the name of the process}

\item{domain}{the domain where the process is defined}

\item{...}{parameters required to define the process
\describe{
  \item{dispersion}{required by Wiener process}
  \item{sigma}{required by white noise process}
  \item{eigen.values}{required by process defined via K-L representation}
  \item{eigen.functions}{required by process defined via K-L representation}
  \item{distribution}{the distribution of PC scores, required by process defined via K-L representation}
}}
}
\value{
a function hanlde in the form of \code{X(tObs,n)} which generates \code{n} independent trajectories observed at tObs
}
\description{
create a centered random process objecct
}
\examples{
X <- centered.process(name='wiener',dispersion=1)
X(regular.grid(50),25)
}
