% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processes.R
\name{matern}
\alias{matern}
\title{The Matern covariance function}
\usage{
matern(x, y = NULL, nu = 1, rho = 1, sig = 1)
}
\arguments{
\item{x}{a vector}

\item{y}{If NULL, then y=x.}

\item{nu}{see https://en.wikipedia.org/wiki/Mat?rn_covariance_function}

\item{rho}{see https://en.wikipedia.org/wiki/Mat?rn_covariance_function}

\item{sig}{see https://en.wikipedia.org/wiki/Mat?rn_covariance_function}
}
\value{
A matrix M of len(X)*len(Y) is returned, where M(i,j)=C(X(i),Y(j))
}
\description{
The Matern covariance function
}
