% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dense.fd.R
\name{plot.dense.fd}
\alias{plot.dense.fd}
\title{plot densely and regularly observed data}
\usage{
\method{plot}{dense.fd}(y, ...)
}
\arguments{
\item{y}{the dense data object generated from \code{gen.dense.fda}}

\item{...}{other parameters passed to \code{matplot}}
}
\description{
plot densely and regularly observed data
}
