% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irreg.fd.R
\name{plot.sparse.fd}
\alias{plot.sparse.fd}
\title{Plot Irregular Functional Data}
\usage{
\method{plot}{sparse.fd}(x, ...)
}
\arguments{
\item{x}{the object generated by \code{\link{irreg.fd}}.}

\item{...}{other parameters passed to \code{plot} and \code{lines}.}
}
\value{
a plot of the dataset \code{x}.
}
\description{
Plot Irregular Functional Data
}
