% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse.fd.R
\name{plot.sparse.fd}
\alias{plot.sparse.fd}
\title{plot sparse functional data}
\usage{
\method{plot}{sparse.fd}(R, ...)
}
\arguments{
\item{R}{the object generated by \code{gen.sparse.fda}}

\item{...}{other parameters passed to \code{plot} and \code{lines}}
}
\description{
plot sparse functional data
}
