% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{regular.grid}
\alias{regular.grid}
\title{generate a regular grid of points on a given interval}
\usage{
regular.grid(M = 100, domain = c(0, 1), h = 1/(2 * M))
}
\arguments{
\item{M}{the number of points}

\item{domain}{the interval}

\item{h}{margin at boundaries}
}
\description{
generate a regular grid of points on a given interval
}
