% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processes.R
\name{white.noise}
\alias{white.noise}
\title{Create a White Noise Process}
\usage{
white.noise(sig = 1)
}
\arguments{
\item{sig}{sigma parameter of the white noise process}
}
\value{
a function hanlde in the form of \code{X(tObs,n)} which generates \code{n} independent trajectories observed at \code{tObs}.
}
\description{
Create a White Noise Process
}
\examples{
X <- white.noise()
X(regular.grid(50),25)
}
